+++ compiler/src/test/scala/dx/translator/TranslatorTest.scala
@@ -1566,4 +1566,10 @@ Main.compile(args.toVector) shouldBe a[SuccessfulCompileIR]
     val args = path.toString :: cFlags
     Main.compile(args.toVector) shouldBe a[SuccessfulCompileIR]
   }
+
+  it should "translate a workflow with references between output variables" in {
+    val path = pathFromBasename("draft2", "output_references.wdl")
+    val args = path.toString :: cFlags
+    Main.compile(args.toVector) shouldBe a[SuccessfulCompileIR]
+  }
 }