+++ core/src/main/scala/dx/core/languages/wdl/WdlUtils.scala
@@ -711,8 +711,8 @@ object WdlUtils {
   }
 
   /**
-    * A trivial expression has no operators, it is either(1) a constant,
-    * (2) a single identifier, or (3) an access to a call field.
+    * A trivial expression has no operators and it is either a constant
+    * or a single identifier or an access to a call field.
     * For example, `5`, `['a', 'b', 'c']`, and `true` are trivial.
     * 'x + y'  is not.
     * @param expr expression
@@ -720,9 +720,8 @@ object WdlUtils {
     */
   def isTrivialExpression(expr: TAT.Expr): Boolean = {
     expr match {
-      case expr if TypeUtils.isPrimitiveValue(expr) => true
-      case _: TAT.ExprIdentifier                    => true
-
+      case _ if TypeUtils.isPrimitiveValue(expr) => true
+      case _: TAT.ExprIdentifier                 => true
       // A collection of constants
       case TAT.ExprPair(l, r, _)   => Vector(l, r).forall(TypeUtils.isPrimitiveValue)
       case TAT.ExprArray(value, _) => value.forall(TypeUtils.isPrimitiveValue)
@@ -731,15 +730,13 @@ object WdlUtils {
           case (k, v) => TypeUtils.isPrimitiveValue(k) && TypeUtils.isPrimitiveValue(v)
         }
       case TAT.ExprObject(value, _) => value.values.forall(TypeUtils.isPrimitiveValue)
-
       // Access a field in a call
       //   Int z = eliminateDuplicate.fields
       // TODO: this will work for structs as well, if we are able to make
       //  struct fields part of the output closure (see comment in
       //  getClosureInputsAndOutputs)
       case TAT.ExprGetName(TAT.ExprIdentifier(_, _: T_Call), _, _) => true
-
-      case _ => false
+      case _                                                       => false
     }
   }
 
@@ -1044,11 +1041,14 @@ object WdlUtils {
     */
   def getOutputClosure(outputs: Vector[TAT.OutputParameter]): Map[String, T] = {
     // create inputs from all the expressions that go into outputs
+    // exclude any inputs that are references to other output parameters
+    val paramNames = outputs.map(_.name).toSet
     outputs
       .flatMap {
         case TAT.OutputParameter(_, _, expr) => Vector(expr)
       }
       .flatMap(e => getExpressionInputs(e, withField = false))
+      .filterNot(i => paramNames.contains(i.identifier))
       .groupBy(_.fullyQualifiedName)
       .map {
         // if there are multiple references to the same parameter, make sure the