+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 ## in develop
 
 * Fixes issue where an optional variable inside a conditional could have an inferred type of `T??`, which is illegal and results in a runtime error
+* Fixes issue with references between output variables
+* Uses an output stage if there is an output parameter with a literal value (DNAnexus output parameters do not support default values)
 
 ## 2.4.7 2021-06-09